separated deletion of vertices into separate funciton
The deletion of vertices is a bit complicated,  because we dont want to delete a vertex if it is used by other geometry such as a polygon.
Separating it out allows us to delete vertices easily from within other functions, in stead of only selecting them via the GUI.
This will come in handy for the rack_bay creation and update functionality.